+++ .storybook/config.js
@@ -1,7 +1,11 @@
-import { configure } from '@storybook/react';
+import { configure, addDecorator } from '@storybook/react';
+import centered from '@storybook/addon-centered';
+
+addDecorator(centered);
 
 function loadStories() {
-  require('../src/stories');
+  require('../src/components/Square/stories')
+  require('../src/components/Board/stories')
 }
 
 configure(loadStories, module);