+++ src/components/Board/stories.js
@@ -1,0 +1,7 @@
+import React from 'react';
+import { storiesOf } from '@storybook/react';
+
+import { Board } from '..';
+
+storiesOf('Board', module)
+  .add('empty', () => <Board />);