+++ src/components/Square/stories.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import { storiesOf } from '@storybook/react';
+
+import { Square } from '..';
+
+storiesOf('Square', module)
+  .add('black', () => <Square color="black" />)
+  .add('white', () => <Square color="white" />);