+++ src/components/index.js
@@ -1,0 +1,3 @@
+// Keep in alphabetical order
+export { default as Square } from './Square/Square';
+export { default as Board } from './Board/Board';